+++ apps/dot-components-e2e/src/integration/Chip/Chip.spec.ts
@@ -1,5 +1,5 @@
 describe('dot-components: Chip component', () => {
-  before(() => cy.visit('/iframe.html?id=experimental-chip--default'));
+  before(() => cy.visit('/iframe.html?id=components-chip--default'));
 
   it('should have a dot- prefix', () => {
     cy.get('div').should('have.class', 'dot-chip');